Limit calendar views to 1 year, fix events pagination (#278)
Previously we were not paginating results over the events from nationbuilder. This meant that we were only returning a max results of 500, no matter
what the timeframe was.

We never noticed this on the events page since we only return results for that month, which never exceeded the 500 total. However for the
ical value we did not pass in a end date, which meant that we were only returning 500 possibly many more alarms.

Additionally, the results returned by the nationbuilder api are not sorted by the date of the event, but rather the date at which the event
was created, which meant that some events in the near future were not showing up in the results.

Fix is 2 parts:
- First only return the results from the next year within the calendar by default on the ical view
- For all events calls, add in pagination so all results are returned in the proper timeframe